Add test for image.verify()
This test is mostly to satisfy the unused linter. The test should be
expanded to actually test the function. Because of the change of the
internal file type to io.ReadSeeker, it means that we can
programatically create a imaginary file in the test, rather than relying
on actual test files.