Make the provider more idiomatic
The provider was originally written on the old SDK and some of the
patterns were not updated. This more closely follows the pattern from
hashicorp scaffolding provider.

There is no need for the API to be public, hence they are in the
internal directory.

Also included is a bump in go-virtualbox version to include the new
Manager interface which will be called in resources and data sources
improving testability of provider.